minor float conversions

+++ python/prepareSequentialTraining.py
@@ -368,7 +368,7 @@ def pickIndex(var):
 def create_stream(df, ix_train, ix_test, num_obj, sort_col, VAR_FILE_NAME):
   n_variables = df.shape[1]
   var_names = df.keys()
-  data = np.zeros((df.shape[0], num_obj, n_variables), dtype='float32')
+  data = np.zeros((df.shape[0], num_obj, n_variables), dtype='float64')
   
   # call functions to build X (a.k.a. data)
   sort_objects(df, data, sort_col, num_obj)
@@ -402,7 +402,7 @@ def sort_objects(df, data, SORT_COL, max_nobj):
     # objs = [[pt's], [eta's], ...] of particles for each event
 
     #objs = np.array([v.tolist() for v in event.get_values()], dtype='float32')[:, (np.argsort(event[SORT_COL]))[::-1]]
-    objs = np.array([v for v in event.get_values()], dtype='float32')[:, (np.argsort(event[SORT_COL]))[::-1]]
+    objs = np.array([v for v in event.get_values()], dtype='float64')[:, (np.argsort(event[SORT_COL]))[::-1]]
     # total number of tracks per jet      
     nobjs = objs.shape[1] 
     # take all tracks unless there are more than n_tracks
